Server Rendering - created browser module - created server module - moved App.js to browser - moved uiserver.js to server - added .babelrc to browser - added .eslint to server and browser
