+++ ui/browser/.babelrc
@@ -1,0 +1,14 @@
+{
+ "presets": [
+ ["@babel/preset-env", {
+ "targets": {
+ "ie": "11",
+ "edge": "15",
+ "safari": "10",
+ "firefox": "50",
+ "chrome": "49"
+ }
+ }],
+ "@babel/preset-react"
+ ]
+}
