+++ ui/browser/.eslintrc
@@ -1,0 +1,12 @@
+{
+    "env": {
+        "browser": true
+    },
+    "rules": {
+        "import/extensions": [ "error", "always",
+            { "ignorePackages": true
+            }
+        ],
+        "react/prop-types": "off"
+    }
+}
