+++ ui/server/.eslintrc
@@ -1,0 +1,8 @@
+{
+    "env": {
+        "node": true
+    },
+    "rules": {
+        "no-console": "off"
+    }
+}
