+++ ui/server/uiserver.js
@@ -21,7 +21,7 @@ if (enableHMR && (process.env.NODE_ENV !== 'production')) {
   const devMiddleware = require('webpack-dev-middleware');
   const hotMiddleware = require('webpack-hot-middleware');
 
-  const config = require('./webpack.config.js');
+  const config = require('../webpack.config.js');
   config.entry.app.push('webpack-hot-middleware/client');
   config.plugins = config.plugins || [];
   config.plugins.push(new webpack.HotModuleReplacementPlugin());
@@ -30,7 +30,6 @@ if (enableHMR && (process.env.NODE_ENV !== 'production')) {
   app.use(devMiddleware(compiler));
   app.use(hotMiddleware(compiler));
 }
-
 
 // importing http-proxy-middleware object
 // const proxy = require('http-proxy-middleware');
