+++ ui/src/.eslintrc
@@ -1,0 +1,11 @@
+{
+    "env":{
+        "browser": true,
+        "node": true
+    },
+        "rules":{
+        "import/extensions" : ["error" , "always" , {"ignorePackages" : true }],
+        "no-console": "off",
+        "react/props-types" : "off"
+    }
+}
