+++ ui/webpack.config.js
@@ -4,7 +4,7 @@ const path = require('path');
 
 module.exports = {
   mode: 'development',
-  entry: { app: ['./src/App.jsx'] },
+  entry: { app: ['./browser/App.jsx'] },
   output: {
     filename: '[name].bundle.js',
     path: path.resolve(__dirname, 'public'),
